+++ src/Fci.cpp
@@ -509,7 +509,7 @@ std::vector<Node> Fci::getSepset(const Node& a, const Node& b, SepsetMap& sepset
 
     for (int d = 0; d <= maxDepth; d++) {
 
-	if (d <= ppa.size()) {
+	if (d <= (int) ppa.size()) {
 	    ChoiceGenerator cg1(ppa.size(), d);
 	    std::vector<int> *comb2;
 	    for (comb2 = cg1.next(); comb2 != NULL; comb2 = cg1.next()) {
@@ -523,7 +523,7 @@ std::vector<Node> Fci::getSepset(const Node& a, const Node& b, SepsetMap& sepset
 	    }
 	}
 
-	if (d <= ppb.size()) {
+	if (d <= (int) ppb.size()) {
 	    ChoiceGenerator cg2(ppb.size(), d);
 	    std::vector<int> *comb3;
 	    for (comb3 = cg2.next(); comb3 != NULL; comb3 = cg2.next()) {
@@ -552,7 +552,7 @@ std::vector<Node> Fci::getMinSepset(const Node& a, const Node& b,
     std::vector<Node> minSepset;
     double pval = 0;
     
-    for (int d = 0; d <= sepset.size(); d++) {
+    for (int d = 0; d <= (int) sepset.size(); d++) {
 	
         ChoiceGenerator cg(sepset.size(), d);
 	std::vector<int> *comb;
